+++ studio/src/components/layout/title-layout.tsx
@@ -4,16 +4,18 @@ export const TitleLayout = ({
   title,
   subtitle,
   items,
+  toolbar,
   children,
 }: {
   title: string;
   subtitle: string;
   items?: React.ReactNode;
+  toolbar?: React.ReactNode;
   children?: React.ReactNode;
 }) => {
   return (
-    <>
-      <div className="sticky top-0 z-10 bg-background">
+    <div className="flex h-full flex-col">
+      <div className="sticky top-0 z-10 bg-background md:top-6 lg:top-0">
         <div className="flex flex-col justify-between gap-y-4 px-4 pt-4 lg:flex-row lg:items-center lg:px-6">
           <div className="space-y-2">
             <h1 className="text-2xl font-bold">{title}</h1>
@@ -22,10 +24,9 @@ export const TitleLayout = ({
           {items}
         </div>
         <Separator className="mt-4" />
+        {toolbar}
       </div>
-      <div className="h-[calc(100vh_-_15rem)] flex-1 px-4 py-4 lg:h-[calc(100%_-_98px)] lg:px-6">
-        {children}
-      </div>
-    </>
+      <div className="h-auto flex-1 px-4 py-4 lg:px-6">{children}</div>
+    </div>
   );
 };